+++ client/ui/Credits.jsx
@@ -34,6 +34,7 @@ export default class Credits extends Component {
                       <ul>
                         <b>6.4 </b>
                         <li>Aggregate section updated to allow layer and sublayer filtering in the query</li>
+                        <li>Suggest support for query layer and sublayer filtering added</li>
                       </ul>
 
                       <ul>